Code works. Coefficient is still showing as 0.0

+++ Models/TableModel.java
@@ -3,9 +3,9 @@ package Models;
 public class TableModel {
     public int size;
     public int countAverage;
-    public int countCoef;
+    public double countCoef;
     public long timeAverage;
-    public long timeCoef;
+    public double timeCoef;
 
     public TableModel() {
     }
